requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==7.0.0
anyio==3.6.1
appdirs==1.4.4
appnope==0.1.2
APScheduler==3.9.1
argon2-cffi==21.1.0
async-timeout==4.0.2
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
Brotli==1.0.9
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
dash==2.6.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
dbus-python==1.2.18
debugpy==1.6.2
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.16.1
Flask==2.2.2
Flask-Compress==1.12
fonttools==4.29.1
frozenlist==1.3.1
fs==2.4.12
gast==0.5.2
gyp==0.1
html5lib==1.1
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
ipykernel==6.0.0
ipympl==0.9.1
ipython==7.31.1
ipython_genutils==0.2.0
ipyvue==1.8.0
ipyvuetify==1.8.4
ipywidgets==7.7.1
ipywidgets-toggle-buttons==0.2.3
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.3
joblib==1.2.0
json5==0.9.9
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyter-server==1.18.1
jupyter-server-proxy==3.2.1
jupyterlab==3.4.5
jupyterlab-dash==0.1.0a3
jupyterlab-pygments==0.1.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
jupyterthemes==0.20.0
kiwisolver==1.3.2
lesscpy==0.15.0
lxml==4.8.0
lz4==3.1.3+dfsg
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.10.0
mplcursors==0.5.1
mpmath==0.0.0
multidict==6.0.2
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.30
nbclassic==0.4.3
nbclient==0.5.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.4
notebook==6.4.8
notebook-shim==0.1.0
numpy==1.21.5
olefile==0.46
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.1
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.10.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.20.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pydantic==1.7.4
Pygments==2.4.1
PyGObject==3.42.1
pymannkendall==1.4.2
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.1
pythran==0.10.0
pytz==2020.5
pywatchman==1.4.1
PyYAML==5.4.1
pyzmq==22.3.0
regex==2022.7.25
requests==2.28.1
Rx==1.6.1
SciencePlots==1.0.9
scikit-learn==1.1.2
scipy==1.8.0
seaborn==0.12.1
Send2Trash==1.8.1b0
simpervisor==0.4
six==1.15.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.3.22
statsmodels==0.13.2
stringcase==1.2.0
sympy==1.9
tenacity==8.0.1
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
typing-inspect==0.6.0
tzlocal==2.1
ufoLib2==0.13.1
ujson==4.0.2
unicodedata2==14.0.0
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.2
widgetsnbextension==3.6.1
wrapt==1.12.1
xxhash==2.0.0
yarl==1.8.1
zipp==1.0.0
